fix(tests): add tests for AppState, SoundManager, and ScheduledBellAlert
- Add tests for AppState
- Add tests for SoundManager
- Add tests for ScheduledBellAlert

All tests are passing except for ScheduledBellAlertTests. We don't use
that class anymore, so we'll remove it in the next commit.

+++ LittleMoments/Core/Audio/ScheduledBellAlert.swift
@@ -1,3 +1,4 @@
+import AVFoundation
 import Foundation
 
 protocol ScheduledAlert: Equatable {
@@ -85,7 +86,7 @@ class RecurringScheduledBellAlert: ScheduledAlert {
   func getProgress(secondsElapsed: CGFloat) -> CGFloat {
     let timeUntilNextTrigger = (targetTimeInSec - secondsElapsed).truncatingRemainder(
       dividingBy: intervalInSec)
-    return CGFloat(timeUntilNextTrigger) / CGFloat(intervalInSec)
+    return timeUntilNextTrigger / intervalInSec
   }
 
   func checkTrigger(secondsElapsed: CGFloat) {
